% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brewSider.R
\name{brewSider}
\alias{brewSider}
\title{Simple isotope estimation.}
\usage{
brewSider(species, habitat, tissue, diet.type, isotope, verbose = TRUE)
}
\arguments{
\item{species}{a binomial species name from a mammal or a bird.}

\item{habitat}{a character string of either \code{"marine"} or \code{"terrestrial"}.}

\item{tissue}{the tissue type. Can be one of the following: \code{c("blood", "claws", "collagen", "feather", "hair", "kidney", "liver", "milk", "muscle")}.}

\item{diet.type}{the diet type. Can be one of the following: \code{c("carnivore", "herbivore", "omnivore",  "pellet")}.}

\item{isotope}{the isotope for which discrimination factor is to be imputed,  either \code{"carbon"} or \code{"nitrogen"}.}

\item{verbose}{whether to be verbose or not (\code{default = TRUE}).}
}
\value{
The estimated isotope value.
}
\description{
Runs a simple isotope estimation with the minimum number of parameters.
}
\details{
The estimation is run on a single randomly selected tree with fixed MCMC parameters.
}
\examples{


brewSider(species = "Meles_meles", habitat = "terrestrial", tissue = "blood",
      diet.type = "omnivore", isotope = "carbon")

}
\author{
Thomas Guillerme
}

