% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDER.R
\docType{data}
\name{combined_trees}
\alias{combined_trees}
\alias{combined_tree}
\title{Avian and mammalian trees}
\format{Contains a \code{multiPhylo} object of two randomly combined trees of
  birds and mammals.}
\description{
This is a dataset containing trees from birds (Jetz et al, 2012) and mammals
(Kuhn et al, 2011).
}
\references{
[ADD REFERENCES]
}
\keyword{datasets}
