% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEsiR.R
\docType{package}
\name{DEsiR}
\alias{DEsiR}
\alias{DEsiR-package}
\title{DEsiR: Estimating istopic fractionation / trophic enrichment factors}
\description{
DEsiR uses a phylogenetic regression model to estimate the isotopic 
fractionation or trophic enrichment factor depending on ones preference for 
terminology. Estimates are imputed using Bayesian inference, and draw from 
species in close taxonomic proximity as well as including information on the 
tissue type and ecology of the desired estimate. The database used to
generate these predictions can be updated in the future as more infomraiton
becomes available, which will lead to ever more accurate and precise
estimates. Only mammals and birds are currently included in the pacakge, but
again this can be expanded to other taxonomic groups as more data becomes
available. We provide examples of how to use custom databases and custom
models for prediction.
}
\section{DEsiR functions}{
 The DEsiR functions ...
}

