% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeSider.R
\name{imputeSider}
\alias{imputeSider}
\title{Running MCMCglmm}
\usage{
imputeSider(mulTree.data, formula, random.terms, parameters = c(1200000,
  500, 2e+05), priors, chains = 2, convergence = 1.1, ESS = 1000,
  output = "teff_output", save.model = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{mulTree.data}{output from \code{\link{prepareSider}} containing data and
phylogeny as a \code{mulTree} object.}

\item{formula}{an object of class \code{formula} describing the fixed 
effects.}

\item{random.terms}{Optional. An object of class \code{formula} describing
the random effects. If missing, the formula from
\code{mulTree.data$random.terms} is used.}

\item{parameters}{a list of three numerical values to be used respectively
as: (1) the number of iterations, (2) the sampling value, (3) the burnin.}

\item{priors}{optional list of prior specifications (see details).}

\item{chains}{The number of MCMC for each run (\code{default = 2}).}

\item{convergence}{limits set for the point estimates of the potential scale 
reduction factor (see \code{link[coda]{gelman.diag}} - \code{default =
1.1}).}

\item{ESS}{effective sample size of MCMC iterations for each model estimate
(\code{default = 1000}).}

\item{output}{The name of the output files (\code{default = "teff_output"})}

\item{save.model}{whether to save the models out of R environment (default)
or just get the estimates.}

\item{verbose}{whether to be verbose or not (\code{default = TRUE}).}

\item{...}{any optional arguments to be passed to
\code{\link[mulTree]{mulTree}}.}
}
\value{
list containing Posterior Distributions of imputed discrimination 
  factors including a posterior distribution for each individual chain in 
  Tef.est and a combined chain in Tef.global.
}
\description{
Runs a \code{\link[MCMCglmm]{MCMCglmm}} model to impute delta
}
\details{
The \code{priors} argument must be of 3 possible elements: R 
  (R-structure) G (G-structure) and B (fixed effects). B is a list containing
  the expected value (mu) and a (co)variance matrix (V) representing the 
  strength of belief: the defaults are \code{B$mu = 0} and \code{B$V = 
  I*1e+10}, where where I is an identity matrix of appropriate dimension. The
  priors for the variance structures (R and G) are lists with the expected 
  (co)variances (V) and degree of belief parameter (nu) for the 
  inverse-Wishart, and also the mean vector (alpha.mu) and covariance matrix 
  (alpha.V) for the redundant working parameters. The defaults are \code{nu =
  0}, \code{V = 1}, \code{alpha.mu = 0}, and \code{alpha.V = 0}. When alpha.V
  is non-zero, parameter expanded algorithms are used.
}
\examples{
## Loading data
data(combined_trees); data(isotope_data)

## Setting up the isotope estimation for one species
taxa_estimate <- recipeSider(species = "Meles_meles", habitat = "terrestrial",
     taxonomic.class = "mammalia", tissue = "blood", diet.type = "omnivore", 
     tree = combined_trees)

## Generating the "mulTree" object for the estimation
taxa_est_mulTree <- prepareSider(data.estimate = taxa_estimate,
     data.isotope = isotope_data, tree = combined_trees, isotope = "carbon")

## Setting up the MCMCglmm parameters
MCMC_parameters <- c(1200, 200, 5)
MCMC_formula <- delta13C ~ diet.type + habitat
isotope_estimate <- imputeSider(taxa_est_mulTree, formula = MCMC_formula,
     parameters = MCMC_parameters, save.model = FALSE)

## Print out the results
summary(isotope_estimate$tdf_global)
plot(isotope_estimate$tdf_global)


}
\seealso{
\code{\link{recipeSider}}, \code{\link{prepareSider}},
  \code{\link{combined_trees}}, \code{\link{isotope_data}},
  \code{\link[mulTree]{as.mulTree}}, \code{\link[mulTree]{mulTree}},
  \code{\link[MCMCglmm]{MCMCglmm}}.
}
\author{
Kevin Healy
}
