% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESIR.R
\docType{data}
\name{isotope_data}
\alias{TESIR_data}
\alias{isotope_data}
\title{Isotope data set}
\format{a \code{data.frame} containing isotopic data derived from controlled feeding trials on birds and mammals.
\describe{
       \item{species}{The species binomial name.}
       \item{habitat}{The species habitat (either \code{marine} or \code{terrestrial}).}
       \item{taxonomic.class}{The species taxonomic class (either \code{aves} or \code{mammalia}).}
       \item{tissue}{The measured tissue (\code{liver}, \code{blood}, \code{kidney}, \code{muscle}, \code{hair}, \code{milk}, \code{feather}, \code{claws} or , \code{Collagen}.}
       \item{source.iso.13C}{The isotopic 13C delta of the food source.}
       \item{source.iso.15N}{The isotopic 15N delta of the food source.}
       \item{delta13C}{The tropic discrimination factor for delta 13C.}
       \item{delta15N}{The tropic discrimination factor for delta 15N}
}}
\description{
This is a dataset containing isotope data collected for the \code{TESIR} package.
}
\references{
[ADD REFERENCES]
}
\keyword{datasets}

