% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeSider.R
\name{recipeSider}
\alias{recipeSider}
\title{Specifying observations to be imputed.}
\usage{
recipeSider(species, taxonomic.class, tissue, diet.type, habitat,
  source.iso.13C, source.iso.15N, tree)
}
\arguments{
\item{species}{a binomial species name as a character string.}

\item{taxonomic.class}{the class of species as a character string, either
\code{"mammalia"} or \code{"aves"}.}

\item{tissue}{the tissue type. Can be one of the following: \code{c("blood",
"claws", "collagen", "feather", "hair", "kidney", "liver", "milk",
"muscle")}.}

\item{diet.type}{the diet type. Can be one of the following:
\code{c("carnivore", "herbivore", "omnivore",  "pellet")}.}

\item{habitat}{a character string of either \code{"marine"} or
\code{"terrestrial"}.}

\item{source.iso.13C}{the source carbon isotopic value (must be
\code{numeric}).}

\item{source.iso.15N}{the source nitrogen isotopic value (must be
\code{numeric}).}

\item{tree}{a \code{phylo} or \code{multiPhylo} object.}
}
\value{
A \code{data.frame} object with one row being a checked species for
  imputation.
}
\description{
Specifies which observation to impute.
}
\examples{
## Load the combined trees data
data(combined_trees)

## Initialise the data in the right format
new.data.test <- recipeSider(species = "Meles_meles", habitat = "terrestrial", 
   taxonomic.class = "mammalia", tissue = "blood", diet.type = "omnivore", 
   tree = combined_trees)

}
\seealso{
\code{\link{prepareSider}}, \code{\link{combined_trees}}.
}
\author{
Kevin Healy
}
