% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.isotope_data.R
\name{refresh.isotope_data}
\alias{refresh.isotope_data}
\title{Refreshes the isotope dataset}
\usage{
refresh.isotope_data(data = "SIDER_data.csv")
}
\arguments{
\item{data}{which is typically the SIDER_data.csv file containing the dataset
used for model fitting and imputation.}
}
\value{
Saves the updated file to ../data/isotope_data.rda for use within the
  pacakge
}
\description{
Internal utility function for refreshing the isotope_data 
  dataset (not exported) This function rapidly updates the isotope_data.rda 
  file once one updates the TESIR_data.csv file.
}
\author{
Thomas Guillerme
}
