% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrumpSider.R
\name{scrumpSider}
\alias{scrumpSider}
\title{Loads isotopic and phylogenetic data}
\usage{
scrumpSider(iso.data, tree)
}
\arguments{
\item{iso.data}{A species name (\code{character}, or the \code{numeric} value
of its corresponding row), a vector of species names or \code{"all"} to
display the related isotopic data. If \code{iso.data} is provided then 
\code{tree} should not be provided.}

\item{tree}{The ID of a specific tree to display (\code{numeric}) or a
\code{"all"} to display all trees. If \code{iso.data} is provided, then 
\code{tree} should not be.}
}
\description{
Loads isotopic and phylogenetic data for SIDER analysis and
  allows to visualise data for specific taxa or visualise specific trees.
}
\details{
The function always loads the isotopic data set and the multiple
  phylogenies. Only one or other of \code{iso.data} and \code{tree} should 
  be provided by the user. It will return an error if both are given.
}
\examples{

## Displaying information for a specific species
scrumpSider(iso.data = "Homo_sapiens")

## Displaying information for a specific tree
scrumpSider(tree = 1)


}
\seealso{
\code{link{isotope_data}}, \code{link{combined_trees}}
}
\author{
Thomas Guillerme
}
