% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTefEst.R
\name{setTefEst}
\alias{setTefEst}
\title{Specify the observation to be imputed.}
\usage{
setTefEst(species = c(), habitat = c("marine", "terrestrial", NA),
  taxonomic.class = c("mammalia", "aves", NA), tissue = c("blood", "claws",
  "collagen", "feather", "hair", "kidney", "liver", "milk", "muscle", NA),
  diet.type = c("carnivore", "herbivore", "omnivore", "pellet", NA),
  source.iso.13C = c(), source.iso.15N = c(), phylogeny = c())
}
\arguments{
\item{species}{a binomial species name as a character string.}

\item{habitat}{a character string of either "marine" or "terrestrial"}

\item{taxonomic.class}{the class of species as a character string, either "mammalia" or
"aves"}

\item{tissue}{the tissue type as a character string taking one of
\code{c("blood", "claws", "collagen", "ceather", "hair", "kidney", "liver",
"milk", "muscle")}}

\item{diet.type}{the diet type as a character string taking one of
\code{c("carnivore", "herbivore", "omnivore",  "pellet")}}

\item{phylogeny}{The full phylogenetic tree or set of trees as a
(multi-)phylo object.}

\item{iso_13C}{source carbon isotopic value is available numeric}

\item{iso_15N}{source nitrogen isotopic value is available numeric}
}
\value{
A dataframe object of row length 1 comprising a checked species for
  imputation.
}
\description{
Specify the observation to be imputed.
}

