% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDER.R
\docType{package}
\name{SIDER}
\alias{SIDER}
\alias{SIDER-package}
\title{SIDER: Estimating isotopic fractionation / trophic discrimination factors}
\description{
SIDER uses a phylogenetic regression model to estimate the isotopic 
fractionation or trophic enrichment factor depending on ones preference for 
terminology. Estimates are imputed using Bayesian inference, and draw from 
species in close taxonomic proximity as well as including information on the 
tissue type and ecology of the SIDER-ed estimate. The database used to
generate these predictions can be updated in the future as more information
becomes available, which will lead to ever more accurate and precise
estimates. Only mammals and birds are currently included in the package, but
again this can be expanded to other taxonomic groups as more data becomes
available. We provide examples of how to use custom databases and custom
models for prediction.
}

