% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tefMcmcglmm.R
\name{tefMcmcglmm}
\alias{tefMcmcglmm}
\title{Run the MCMCglmm models to imput delta.}
\usage{
tefMcmcglmm(mulTree.data, formula = delta13C ~ source.iso.15N + diet.type +
  habitat, random.terms = ~animal + species + tissue, nitt = c(12000),
  thin = c(10), burnin = c(2000), prior = NULL, no.chains = c(2),
  convergence = c(1.1), ESS = c(1000))
}
\arguments{
\item{mulTree.data}{output from tefMulClean containing data and phylogeny 
as a mulTree object.}

\item{formula}{an object of class "formula" decribing the fixed effects.}

\item{random.terms}{an object of class "formula" decribing the random effects.}

\item{nitt}{number of MCMC iterations}

\item{thin}{thinning interval of MCMC chain}

\item{burnin}{number of iterations to discard at begining of chain}

\item{prior}{optional list of prior specifications having 3 possible elements: R (R-structure) G (G-structure)
and B (fixed effects). B is a list containing the expected value (mu) and a (co)variance matrix (V) representing
the strength of belief: the defaults are B$mu=0 and B$V=I*1e+10, where where I is an identity matrix of 
appropriate dimension. The priors for the variance structures (R and G) are lists with the expected 
(co)variances (V) and degree of belief parameter (nu) for the inverse-Wishart, and also the mean vector 
(alpha.mu) and covariance matrix (alpha.V) for the redundant working parameters. The defaults are 
nu=0, V=1, alpha.mu=0, and alpha.V=0. When alpha.V is non-zero, parameter expanded algorithms 
are used.}

\item{no.chains}{The number of MCMC chains for each run.}

\item{convergence}{limits set for the point estimates of the potential scale reduction factor
see gelman.diag {coda}.
ESS effective sample size of MCMC iterations for each model estimate.}
}
\value{
list containing Posterior Distributions of imputed discrimination factors  
including a posterior distribution for each indavidual chain in Tef.est and
a combined chain in Tef.global.
}
\description{
Run the MCMCglmm models to imput delta.
}

