% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tefMulClean.R
\name{tefMulClean}
\alias{tefMulClean}
\title{Matches phylogenies to data and creates multree object.}
\usage{
tefMulClean(new.data = c(), data = data, species_col_name = c("species"),
  trees, isotope = c("carbon", "nitrogen"), taxonomic.class = c("mammalia",
  "aves"), random = ~animal + species + tissue)
}
\arguments{
\item{new.data}{data for species to be imputed see setTefEst}

\item{data}{data.frame object matching phylogeny}

\item{species_col_name}{name of column contining binomial species names}

\item{trees}{phylogeny of phylo/multiphylo class}

\item{isotope}{for which discrimination factor is to be imputed}

\item{taxonomic.class}{the class of species as a character string, either 
"mammalia" or "aves"}

\item{random.terms}{an object of class "formula" decribing the random effects.}

\item{phy}{phylogeny of phylo/multiphylo class matching data}
}
\value{
multree object
}
\description{
Matches phylogenies to data and creates multree object.
}

