% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrumpSider.R
\name{scrumpSider}
\alias{scrumpSider}
\title{Loads isotopic and phylogenetic data}
\usage{
scrumpSider(iso.data, tree)
}
\arguments{
\item{iso.data}{A species name (\code{character}, or the \code{numeric} value of its corresponding row), a vector of species names or \code{"all"} to display the related isotopic data.}

\item{tree}{The ID of a specific tree to display (\code{numeric}) or a \code{"all"} to display all trees.}
}
\description{
Loads isotopic and phylogenetic data for SIDER analysis and allows to visualise data for specific taxa or visualise specific trees.
}
\details{
The function always loads the isotopic data set and the multiple phylogenies. When the two parameters \code{data} or \code{tree} have arguments, the function also prints the details of the arguments.
}
\examples{
## Loading the datasets
scrumpSider()

## Displaying information for a specific species
scrumpSider(iso.data = "Homo_sapiens")

## Displaying information for a specific tree
scrumpSider(tree = 1)

## Displaying information for multiple species and trees
scrumpSider(iso.data = c("Homo_sapiens", "Gallus_gallus"),
            tree = c(1,2))

}
\seealso{
\code{link{isotope_data}}, \code{link{combined_trees}}
}
\author{
Thomas Guillerme
}
