% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setTefEst.R
\name{setTefEst}
\alias{setTefEst}
\title{Specify the observation to be imputed.}
\usage{
setTefEst(species = c(), habitat = c("marine", "terrestrial", NA),
  Class = c("mammalia", "aves", NA), tissue = c("blood", "claws",
  "collagen", "ceather", "hair", "kidney", "liver", "milk", "muscle", NA),
  diet_type = c("carnivore", "herbivore", "omnivore", "pellet", NA),
  iso_13C = c(NA), iso_15N = c(NA), phylogeny = c())
}
\arguments{
\item{species}{a binomial species name as a character string.}

\item{habitat}{a character string of either "marine" or "terrestrial"}

\item{Class}{the class of species as a character string, either "mammalia" or "aves"}

\item{tissue}{the tissue type as a character string taking one of \code{c("blood", "claws", "collagen", "ceather", "hair", "kidney", "liver", "milk", "muscle")}}

\item{diet_type}{the diet type as a character string taking one of \code{c("carnivore", "herbivore", "omnivore",  "pellet")}}

\item{iso_13C}{? WE DONT ACTUALLY WANT THIS?}

\item{iso_15N}{? WE DONT ACTUALLY WANT THIS?}

\item{phylogeny}{The full phylogenetic tree or set of trees as a (multi-)phylo object.}
}
\value{
A dataframe object of row length 1 comprising a checked species for imputation.
}
\description{
Specify the observation to be imputed.
}

