% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTefEst.R
\name{setTefEst}
\alias{setTefEst}
\title{Specifying observations to be imputed.}
\usage{
setTefEst(species = c(), habitat = c("marine", "terrestrial", NA),
  taxonomic.class = c("mammalia", "aves", NA), tissue = c("blood", "claws",
  "collagen", "feather", "hair", "kidney", "liver", "milk", "muscle", NA),
  diet.type = c("carnivore", "herbivore", "omnivore", "pellet", NA),
  source.iso.13C = c(), source.iso.15N = c(), tree = c())
}
\arguments{
\item{species}{a binomial species name as a character string.}

\item{habitat}{a character string of either \code{"marine"} or \code{"terrestrial"}.}

\item{taxonomic.class}{the class of species as a character string, either \code{"mammalia"} or \code{"aves"}.}

\item{tissue}{the tissue type. Can be one of the following: \code{c("blood", "claws", "collagen", "ceather", "hair", "kidney", "liver", "milk", "muscle")}.}

\item{diet.type}{the diet type. Can be one of the following: \code{c("carnivore", "herbivore", "omnivore",  "pellet")}.}

\item{source.iso.13C}{the source carbon isotopic value (must be \code{numeric}).}

\item{source.iso.15N}{the source nitrogen isotopic value (must be \code{numeric}).}

\item{tree}{a \code{phylo} or \code{multiPhylo} object.}
}
\value{
A \code{data.frame} object with one row being a checked species for imputation.
}
\description{
Specifies which observation to impute.
}
\author{
Kevin Healy
}

