% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tefMulClean.R
\name{tefMulClean}
\alias{tefMulClean}
\title{Matching phylogenies and data frames}
\usage{
tefMulClean(new.data = c(), data = data, species.col.name = c("species"),
  tree, isotope = c("carbon", "nitrogen"), taxonomic.class = c("mammalia",
  "aves"), random.terms = ~animal + species + tissue)
}
\arguments{
\item{new.data}{data for species to be imputed (see \code{\link{setTefEst}}).}

\item{data}{a \code{data.frame} object matching phylogeny.}

\item{species.col.name}{name of column containing binomial species names.}

\item{tree}{a \code{phylo} or \code{multiPhylo} object.}

\item{isotope}{the isotope for which discrimination factor is to be imputed, either \code{"carbon"} or \code{"nitrogen"}.}

\item{taxonomic.class}{the class of species as a character string, either \code{"mammalia"} or \code{"aves"}.}

\item{random.terms}{an object of class \code{formula} describing the random effects.}
}
\value{
A \code{mulTree} object.
}
\description{
Matches phylogenies and data into a \code{mulTree} object.
}

